Refactorization round 1 completed
