+++ src/App.scss
@@ -279,10 +279,11 @@ h6 {
     margin-right: 2.5rem;
   }
 
-  .tableandcontrols{
+  .tableandaccordions{
     display: flex;
     flex-direction: column;
     width: 100%;
+    align-items: center;
   }
 
   table {
@@ -293,6 +294,7 @@ h6 {
     text-transform: none;
     border-spacing: 3px 2px;
     border-collapse: separate;
+    table-layout: fixed;
     // animation-name: fadeInOut;
     // animation-duration: 2s;
 
@@ -308,6 +310,9 @@ h6 {
         border-radius: 5px;
         transition: all 0.1s ease-in-out;
         z-index: 99;
+        &.descriptionCol{
+          width: 50%;
+        }
 
         &.sticky {
           top: 4px;
@@ -550,7 +555,7 @@ h6 {
   font-weight: 400;
   text-align: center;
   margin-top: 0.2rem;
-  margin-bottom: 0.1rem;
+  margin-bottom: 0.3rem;
   width: 100%;
 
   .react-autosuggest__input {
@@ -3584,7 +3589,7 @@ select {
   }
 
   .mobile-disclaimer-button {
-    background-color: #ffffff;
+    background-color: #fff;
     color: #808080;
     font-size: xx-small;
     justify-content: center;
